Fixes for 0.7 provider implementation (#288)

+++ kernel/kernel-core/src/main/java/io/mosip/kernel/core/bioapi/model/Score.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class Score {
 	
 	private float scaleScore;	  
-	private float internalScore;
+	private long internalScore;
 	private KeyValuePair[] analyticsInfo;
 
 }
